map iteration 1 complete, chaser tweaks
Taking too long mechanic also implemented so the player can't wander for too long. The map has ten possible page spawn locations, and the chaser now spawns more, and can only be driven off with the flashlight. However, it takes a far shorter amount of time to get rid of it.